Migrate train/test methods to a custom finder example
This helps us reduce the finder module to just the essential
options and classes and prunes it from untested and nearly
deprecated code.

We also keep the main API in the finder module clean to just
the reduced set of methods.
